Created lp_data/HighsReformulation.cpp for reformulation methods, and moved (un)doReformulation, handleInfiniteCost and restoreInfiniteCost to it
