+++ highs/Highs.h
@@ -1571,9 +1571,6 @@ class Highs {
 
   void reportModelStats() const;
 
-  HighsStatus doReformulation();
-  void undoReformulation(HighsStatus& optimize_status);
-
   void exactResizeModel() {
     this->model_.lp_.exactResize();
     this->model_.hessian_.exactResize();
@@ -1774,8 +1771,6 @@ class Highs {
   HighsStatus lpKktCheck(const HighsLp& lp, const std::string& message = "");
   HighsStatus invertRequirementError(std::string method_name) const;
 
-  HighsStatus handleInfCost();
-  void restoreInfCost(HighsStatus& optimize_status);
   HighsStatus optionChangeAction();
 
   HighsStatus userScale(HighsUserScaleData& data);
@@ -1803,6 +1798,13 @@ class Highs {
 
   bool tryPdlpCleanup(HighsInt& pdlp_cleanup_iteration_limit,
                       const HighsInfo& presolved_lp_info) const;
+
+  // Reformulation methods
+  HighsStatus doReformulation();
+  void undoReformulation(HighsStatus& optimize_status);
+
+  HighsStatus handleInfiniteCost();
+  void restoreInfiniteCost(HighsStatus& optimize_status);
 };
 
 // Start of deprecated methods not in the Highs class
