+++ highs/lp_data/HStruct.h
@@ -119,7 +119,7 @@ struct HighsLpMods {
   std::vector<double> save_inf_cost_variable_upper;
 
   void clear();
-  void clearInfCostRecord();
+  void clearInfiniteCostRecord();
   void clearSemiVariableRecord();
   bool isClear();
 };
