+++ highs/lp_data/Highs.cpp
@@ -979,11 +979,10 @@ HighsStatus Highs::optimizeHighs() {
   //
   HighsStatus optimize_status = doReformulation();
   if (optimize_status == HighsStatus::kError) return optimize_status;
-  
-  optimize_status =
-    this->multi_linear_objective_.size() ?
-    this->multiobjectiveSolve() :
-    this->optimizeModel();
+
+  optimize_status = this->multi_linear_objective_.size()
+                        ? this->multiobjectiveSolve()
+                        : this->optimizeModel();
 
   undoReformulation(optimize_status);
 
@@ -1096,7 +1095,7 @@ HighsStatus Highs::calledOptimizeModel() {
   // Set undo_mods = false so that returnFromOptimizeModel() doesn't undo any
   // mods that must be preserved - such as when solving a MIP node
   bool undo_mods = false;
-  assert(!model_.lp_.has_infinite_cost_) ;
+  assert(!model_.lp_.has_infinite_cost_);
 
   // Ensure that all vectors in the model have exactly the right size
   exactResizeModel();
@@ -3771,32 +3770,6 @@ void Highs::invalidateRanging() { ranging_.invalidate(); }
 void Highs::invalidateEkk() { ekk_instance_.invalidate(); }
 
 void Highs::clearIis() { iis_.clear(); }
-
-HighsStatus Highs::doReformulation() {
-  HighsStatus status = HighsStatus::kOk;
-  assert(model_.lp_.has_infinite_cost_ ==
-	 model_.lp_.hasInfiniteCost(options_.infinite_cost));
-  if (model_.lp_.has_infinite_cost_) {
-    // If the model has infinite costs, then try to remove them. The
-    // status indicates the success of this operation and, if
-    // it's unsuccessful, the model will not have been modified.
-    status = handleInfCost();
-    if (status != HighsStatus::kOk) {
-      assert(status == HighsStatus::kError);
-      setHighsModelStatusAndClearSolutionAndBasis(HighsModelStatus::kUnknown);
-      return status;
-    }
-    assert(!model_.lp_.has_infinite_cost_);
-  } else {
-    assert(!model_.lp_.hasInfiniteCost(options_.infinite_cost));
-  }
-  return status;
-}
-
-void Highs::undoReformulation(HighsStatus& optimize_status) {
-  this->restoreInfCost(optimize_status);
-  //  this->model_.lp_.unapplyMods();
-}
 
 HighsStatus Highs::completeSolutionFromDiscreteAssignment() {
   // Determine whether the current solution of a MIP is feasible and,
